Resync with latest production code
Somehow the production code got out of sync with the version used to
initialize this repository.  This commit ensures the repo code is in
sync with the latest production code while maintaining the changes
previously applied to remove hard-coded sensitive information from the
source code, as well as to allow MARTI to run from a Docker container.

+++ dice/unsubscribe.php
@@ -1,45 +1,46 @@
-<!DOCTYPE html>
 <html>
-	<head>
-		<title>Unsubscribe from MARTI dice services</title>
-	</head>
-	<body>
-		<?php
-		$email = filter_input( INPUT_POST, "email", FILTER_SANITIZE_EMAIL );
-		if( is_null( $email ) || $email===false ){ $email = filter_input( INPUT_GET, "email", FILTER_SANITIZE_EMAIL ); }
-		if( is_null( $email ) || $email===false ) {
-		// template for first load
-		?>
-			<form method="POST">
-				Enter your email here to unsubscribe your email from MARTI dice services:
-				<br>
-				<input type="text" name="email" width="40">
-				<br>
-				<input type="submit" value="unsubscribe">
-			</form>
-		<?php
-		}
-		else {
-		// template for handling postbacks
-			require_once("dice.class.php");
-			$dice = new dice;
+<head>
+	<title>Unsubscribe from MARTI dice services</title>
+</head>
+<body>
+<?php
+$email = filter_input( INPUT_POST, "email", FILTER_SANITIZE_EMAIL );
+if( is_null( $email ) || $email===false ){ $email = filter_input( INPUT_GET, "email", FILTER_SANITIZE_EMAIL ); }
+if( is_null( $email ) || $email===false ) {
+// template for first load
+?>
+	<form method="post" action="unsubscribe.php">
+		Enter your email here to unsubscribe your email from MARTI dice services:
+		<br/>
+		<input type="text" name="email" width="40" />
+		<br/>
+		<input type="submit" value="unsubscribe">
+	</form>
+<?php
+}
+else {
+// template for handling postbacks
+	require_once("dice.class.php");
+	$dice = new dice;
 
-			if(! $dice->checkIfMailIsRegistered($email)){
-				exit("This email is not registered, or was never validated.");
-			}
+	if( ! dice::checkEmail($email) )
+		exit("Invalid email format! Please enter a valid email address");
 
-			$sql = "DELETE FROM dice_emails WHERE registered_email=?";
+	if(! $dice->checkIfMailIsRegistered($email))
+		exit("This email is not registered, or was never validated.");
 
-			if( $sth = $dice->dbconn->prepare( $sql ) ){
-			  $sth->bind_param('s',$email);
-			  $sth->execute() or trigger_error($mysqli->error);
-			} else {
-			  echo "A DB error has occured, please contact an admin. (2-"; var_dump( $dice->dbconn->errno );
-			  echo ")";
-			  exit;
-			}
-			echo "Your email was successfully removed. You will no longer receive dice emails.";
-		}
-		?>
-	</body>
+	$sql = "DELETE FROM dice_emails WHERE registered_email=?";
+
+if( $sth = $dice->dbconn->prepare( $sql ) ){
+  $sth->bind_param('s',$email);
+  $sth->execute() or trigger_error($mysqli->error);
+} else {
+  echo "A DB error has occured, please contact an admin. (2-"; var_dump( $dice->dbconn->errno ); echo ")";exit;
+}
+
+	echo "Your email was successfully removed. You will no longer receive dice emails.";
+}
+?>
+
+</body>
 </html>
